chore: moving example file to specific folder
