+++ example/dc_simloop_config.yaml
@@ -1,0 +1,3 @@
+Dc:
+  name: dc
+  type: DcFilterDouble
