+++ include/rosneuro_filters_dc/Dc.hpp
@@ -6,36 +6,34 @@
 #include <gtest/gtest_prod.h>
   
 namespace rosneuro {
+    template <typename T>
+    class Dc : public Filter<T> {
+        public:
+            Dc(void);
+            ~Dc(void) {};
 
-template <typename T>
-class Dc : public Filter<T> {
-	public:
-		Dc(void);
-		~Dc(void) {};
+            bool configure(void);
+            DynamicMatrix<T> apply(const DynamicMatrix<T>& in);
 
-		bool configure(void);
-		DynamicMatrix<T> apply(const DynamicMatrix<T>& in);
+        private:
+            FRIEND_TEST(DcTestSuite, Constructor);
 
-    FRIEND_TEST(DcTestSuite, Constructor);
+    };
 
-};
+    template<typename T>
+    Dc<T>::Dc(void) {
+        this->name_ = "dc";
+    }
 
-template<typename T>
-Dc<T>::Dc(void) {
-	this->name_ = "dc";
-}
+    template<typename T>
+    bool Dc<T>::configure(void) {
+        return true;
+    }
 
-template<typename T>
-bool Dc<T>::configure(void) {
-	return true;
-}
-
-template<typename T>
-DynamicMatrix<T> Dc<T>::apply(const DynamicMatrix<T>& in) {
-	return in - (in.colwise().mean()).replicate(in.rows(), 1);
-}
-
-
+    template<typename T>
+    DynamicMatrix<T> Dc<T>::apply(const DynamicMatrix<T>& in) {
+        return in - (in.colwise().mean()).replicate(in.rows(), 1);
+    }
 }
 
 #endif
