--- test/dc_simloop_config.launch
@@ -1,11 +1,0 @@
-<launch>	
-	<arg name="framesize" default="32"/>
-	<arg name="datapath" default="$(find rosneuro_filters_dc)"/>
-
-	<rosparam command="load" file="$(find rosneuro_filters_dc)/test/dc_simloop_config.yaml"/>
-	<node name="dc_simloop_config" pkg="rosneuro_filters_dc" type="dc_simloop_config" output="screen">
-		<rosparam param="datapath"  subst_value="True">$(arg datapath)</rosparam>
-		<rosparam param="framesize" subst_value="True">$(arg framesize)</rosparam>
-	</node>
-	
-</launch>
