--- test/dc_simloop_config.yaml
@@ -1,3 +1,0 @@
-Dc:
-  name: dc
-  type: DcFilterDouble
